netty-pipe/src/gash/router/container/MessageContainer.java
/**
 * Copyright 2012 Gash.
 *
 * This file and intellectual content is protected under the Apache License, version 2.0
 * (the "License"); you may not use this file except in compliance with the
 * License.  You may obtain a copy of the License at:
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
 * License for the specific language governing permissions and limitations
 * under the License.
 */
package gash.router.container;

import javax.xml.bind.annotation.XmlAccessType;
import javax.xml.bind.annotation.XmlAccessorType;
import javax.xml.bind.annotation.XmlElement;
import javax.xml.bind.annotation.XmlRootElement;

/**
 * The message container is the frame for message requests. The main message
 * content is held in body. This allows the routing (path) to know how to parse
 * the content.
 * 
 * Note encoding of this message using the RoutingDecoderPipeline requires
 * messages to not contain newlines within the message - this is restricting
 * though for demonstration/prototype is is an acceptable req. However, a req
 * that is problematic is all byte data must be encoded (e.g. base64)
 * 
 * @author gash
 * 
 */
@XmlRootElement(name = "message")
@XmlAccessorType(XmlAccessType.PROPERTY)
public class MessageContainer {
	public int getStatusCode() {
		return statusCode;
	}

	public void setStatusCode(int statusCode) {
		this.statusCode = statusCode;
	}

	public String getStatusMsg() {
		return statusMsg;
	}

	public void setStatusMsg(String statusMsg) {
		this.statusMsg = statusMsg;
	}

	private long msgid;
	private String path;
	private String body;

	// response provided (1 = success, 0 = failure)
	@XmlElement(nillable = true)
	private int statusCode = -1;

	@XmlElement(nillable = true)
	private String statusMsg;

	
	public MessageContainer() {
	}

	public String getPath() {
		return path;
	}

	public void setPath(String path) {
		this.path = path;
	}

	public String getBody() {
		return body;
	}

	public void setBody(String body) {
		this.body = body;
	}

	public long getMsgid() {
		return msgid;
	}

	public void setMsgid(long msgid) {
		this.msgid = msgid;
	}
}
